requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyrogram==2.0.106
tgcrypto==1.2.5
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0 